adding basic Controller advice with basic error handling

+++ src/main/java/com/example/producttestapi/service/VoucherServiceImpl.java
@@ -33,6 +33,9 @@ public class VoucherServiceImpl implements VoucherService {
 
     @Override
     public void deleteVoucher(Long id) {
+        if (!voucherRepo.existsById(id)) {
+            throw new ResourceNotFoundException("Voucher not found with id: " + id);
+        }
         voucherRepo.deleteById(id);
     }
 }
